requirements.txt
langchain==0.1.11
youtube-transcript-api
openai
tiktoken
python-dotenv
gradio==4.20.1

